ai-service/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# OCR & Document Processing
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
httpx>=0.27.0

# Machine Learning & NLP
scikit-learn>=1.5.0
numpy>=1.26.0
sentence-transformers>=3.0.0

# Google Gemini API
google-genai>=1.10.0

# Text-to-Speech
gTTS>=2.5.0

# Utilities
python-multipart>=0.0.17
